% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunksum.R
\name{chunk.sum}
\alias{chunk.sum}
\title{Chunk.Sum function}
\usage{
chunk.sum(v, n, na.rm = TRUE)
}
\arguments{
\item{v}{Numeric Vector}

\item{n}{Size of chunk}

\item{na.rm}{Remove NAs (default=TRUE)}
}
\description{
Rolling Sum over distinct chunks
}
