% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TEAM.R
\name{create.boxes2D}
\alias{create.boxes2D}
\title{create.boxes2D}
\usage{
create.boxes2D(bin_info, lim, xll, xul, yll, yul)
}
\arguments{
\item{bin_info}{partition}

\item{lim}{default limit}

\item{xll}{user-specified lower x limit}

\item{xul}{user-specified upper x limit}

\item{yll}{user-specified lower y limit}

\item{yul}{user-specified upper y limit}
}
\value{
A \code{\link{matrix}} containing the rectangular coordinates of
each box
}
\description{
create.boxes2D
}
