% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TEAM.R
\name{create.xy.breaks}
\alias{create.xy.breaks}
\title{create.xy.breaks}
\usage{
create.xy.breaks(dat, m, first.axis)
}
\arguments{
\item{dat}{input \code{\link{data.frame}} of observations}

\item{m}{number of bins}

\item{first.axis}{first axis to partition}
}
\description{
create.xy.breaks
}
