% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estchat.R
\name{est.c.hat}
\alias{est.c.hat}
\title{Estimate threshold function}
\usage{
est.c.hat(l, n, theta0, x.l, c.hats, alpha, m.l)
}
\arguments{
\item{l}{Layer}

\item{n}{Number of pooled case and control observations in each layer 1 bin}

\item{theta0}{Nominal boundary level for binomial parameter at layer 1}

\item{x.l}{Vector of case counts in each bin}

\item{c.hats}{Previous c.hats calculated from layers 1 to l-1}

\item{alpha}{Nominal FDR level}

\item{m.l}{Number of leaf hypotheses at layer l}
}
\description{
Estimate threshold to control FDR in multiple testing procedure
}
