% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estFDRhatl.R
\name{est.FDR.hat.l}
\alias{est.FDR.hat.l}
\title{Calculate FDR}
\usage{
est.FDR.hat.l(min.x, max.x, c.prev, n.l, x.l, theta0, l)
}
\arguments{
\item{min.x}{lower limit of searching range}

\item{max.x}{upper limit of searching range}

\item{c.prev}{Previous c.hat from layer l-1}

\item{n.l}{Vector of number of pooled observations in layer l bins}

\item{x.l}{Vector of case counts in each bin at layer l}

\item{theta0}{Nominal boundary level for binomial parameter at layer 1}

\item{l}{Layer}
}
\description{
Step-down multiple-testing procedure
}
