% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expandmat.R
\name{expand.mat}
\alias{expand.mat}
\title{Enumerate binomial support}
\usage{
expand.mat(mat, vec)
}
\arguments{
\item{mat}{Matrix}

\item{vec}{Numeric Vector}
}
\description{
Enumerate possible counts for calculating binomial probability
}
