% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TEAM.R
\name{matsplitter}
\alias{matsplitter}
\title{matsplitter}
\usage{
matsplitter(M, r, c)
}
\arguments{
\item{M}{matrix}

\item{r}{number of rows to split}

\item{c}{number of columns to split}
}
\description{
matsplitter
}
