% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitNoOverlap.R
\name{splitNoOverlap}
\alias{splitNoOverlap}
\title{splitNoOverlap function}
\usage{
splitNoOverlap(vec, seg.length)
}
\arguments{
\item{vec}{Numeric Vector}

\item{seg.length}{Number of distinct chunks to split vec}
}
\description{
Split a vector into distinct chunks of specified size
}
