% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validcounts.R
\name{valid.counts}
\alias{valid.counts}
\title{Valid counts}
\usage{
valid.counts(x, c.prev)
}
\arguments{
\item{x}{Vector}

\item{c.prev}{Calculated chat from layer l-1}
}
\description{
Enumerate matrix of valid counts for a vector of values
}
