% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TEAM.R
\name{TEAM}
\alias{TEAM}
\title{Testing on an Aggregation Tree Method}
\usage{
TEAM(partition_info, alpha, L)
}
\arguments{
\item{partition_info}{Partition for first layer of aggregation tree}

\item{alpha}{Target false discovery rate (FDR) level}

\item{L}{Number of layers in the aggregation tree}
}
\value{
A \code{\link{list}} containing the number of pooled observations in
each bin (n), the number of bins/leaves at each layer (m.l), the indices of
discovered leaves/layer 1 bins (S.list) in each layer and the estimated
layer-specific thresholds (c.hats)
}
\description{
This function performs multiple testing embedded in an
aggregation tree structure in order to identify local differences between two
probability density functions
}
\examples{
## Example with 1D pdfs: find where case density is higher than control
# density
require(ks)
set.seed(1)
# Simulate local shift difference for each sample from mixture of normals
# Uses rnorm.mixt from ks library
N1 <- N2 <- 1e6
#Controls
x1 <- rnorm.mixt(N1,mus=c(0.2,0.89),sigmas=c(0.04,0.01),props=c(0.97,0.03))
#Cases
x2 <- rnorm.mixt(N2,mus=c(0.2,0.88),sigmas=c(0.04,0.01),props=c(0.97,0.03))
#Create partition
#Create data frames for data
x1.df <- data.frame(X=x1)
x2.df <- data.frame(X=x2)
#Choose number of bins
m = 2^14 #Close to ((N1+N2))^(2/3)
partition1D <- create_partition_info(df1=x1.df,df2=x2.df,m=m)
res <- TEAM(partition_info=partition1D,alpha=0.05,L=3)
#Indices of layer 1 bins rejected at each aggregation layer (not run)
#res$S.list
#Map rejected bins to corresponding regions (not run)
#partition1D$bin.df$x.bd[res$S.list[[3]]]
}
\references{
Pura J, Li X, Chan C, Xie J. TEAM: A Multiple Testing Algorithm
on the Aggregation Tree for Flow Cytometry Analysis
\url{https://arxiv.org/abs/1906.07757}
}
