% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TEAM.R
\name{valid.counts}
\alias{valid.counts}
\title{valid.counts}
\usage{
valid.counts(x, c.prev)
}
\arguments{
\item{x}{count}

\item{c.prev}{count that controls FDR at previous layer l-1}
}
\description{
valid.counts
}
